Complete the tool: Swift tokeniser, sampling, chat — and fix an FP16 overflow
Godwit no longer needs Python. A byte-level BPE tokeniser reads the
checkpoint's own tokenizer.json, so there is no second source of truth to drift
from the weights, and it matches tiktoken exactly on fourteen cases including
emoji with ZWJ sequences, CJK, Cyrillic, mixed case with contractions, and the
harmony special tokens.

Also adds temperature, top-k, top-p and repetition penalty with a seeded
splitmix64 generator; a Conversation type that builds harmony prompts and
separates the model's reasoning channel from its answer; and `godwit chat`,
interactive or one-shot, streaming as it goes.

The important fix is an FP16 overflow. Some prompts produced nothing but token
0 forever -- "What is 17 times 23?" reliably, "What is the capital of France?"
never. The logits were NaN. GPT-OSS was trained in bfloat16, which reaches
~3e38; FP16 stops at 65504, and certain inputs drove the residual stream past
it. The stream is now FP32, with normalised values still handed to the kernels
as FP16 since RMSNorm bounds them.

Worth noting how this was nearly missed. The regression tests all passed
throughout, because they check one layer against a fixed input that never
overflows. What caught it was running the model on a new prompt and reading the
output -- and I had stopped doing that during the optimisation work, checking
only the timing lines. Two commits shipped with this latent.

Also fixed the streamed display, which spliced fragments like
"<|channel|>analysis1\)" together. It diffed by length against a string that
shortens when a channel marker resolves mid-stream; it now only appends when
the new text genuinely extends what was printed, and Conversation.split returns
empty rather than handing back a half-written marker as content.

One test failure here was the test's fault, not the code's: the repetition
penalty case used values where the penalty produced an exact tie, which proves
nothing either way.

+++ Sources/Godwit/Benchmark/LayerTrace.swift
@@ -133,7 +133,7 @@ public struct LayerTrace {
     }
 
     /// Embeddings for a spread of token ids, dequantised from the trunk.
-    private func initialHidden(tokenCount: Int) throws -> [Float16] {
+    private func initialHidden(tokenCount: Int) throws -> [Float] {
         let spec = reader.manifest.spec
         let width = spec.hiddenSize
         let embed = try reader.loadTrunk(section: "embed", device: context.device)
@@ -143,7 +143,7 @@ public struct LayerTrace {
 
         let codes = embed.contents()
         let meta = codes.advanced(by: rows * width)
-        var out = [Float16](repeating: 0, count: tokenCount * width)
+        var out = [Float](repeating: 0, count: tokenCount * width)
 
         let stride = max(1, spec.vocabularySize / max(tokenCount, 1))
         for token in 0..<tokenCount {
@@ -157,7 +157,7 @@ public struct LayerTrace {
                 let zero = BFloat16.toFloat(metaRow[group * 2 + 1])
                 for i in 0..<Int8Affine.groupSize {
                     let index = group * Int8Affine.groupSize + i
-                    out[token * width + index] = Float16(Float(codeRow[index]) * scale + zero)
+                    out[token * width + index] = Float(codeRow[index]) * scale + zero
                 }
             }
         }
